+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All Notable changes to `Spider` will be documented in this file
 - First Party Driver for OrientDB (v2.*)
 - Register configs inside Connections\Manager
 - Define return object from Connections\Manager or Connection
+- Connection Manager caches and can return already instantiated connection
 
 ## v0.1.1 - 4-28-2015
 ### Added